backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
fastapi==0.135.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
gTTS==2.5.4
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
openai==2.28.0
opencv-python==4.13.0.92
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pi_heif==1.3.0
pillow==12.1.1
platformdirs==4.9.4
polars==1.39.0
polars-runtime-32==1.39.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tokenizers==0.22.2
torch==2.12.0.dev20260314
torchaudio==2.11.0.dev20260314
torchvision==0.26.0.dev20260314
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
typer==0.12.5
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
